requirements.txt
PyYAML
beautifulsoup4
requests
python-dateutil
pytz
loguru

